refactor: fix cli
